# backend/openapi.yml
openapi: 3.1.0
info:
  title: Fahrtenbuch API
  version: 1.0.0-dev
  description: |
    REST-API für das selbst gehostete Fahrtenbuch. Die API umfasst
    Authentifizierung, MFA, Passkeys, App-Pairing, Geräteverwaltung,
    Benutzer, Fahrzeuge, Fahrten, Tracking, Statistiken, Import/Export,
    Geocoding sowie zentrale Kartendienst-Konfiguration.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
    description: Relativ zum Fahrtenbuch-Server
security:
  - bearerAuth: []
...
paths:
...
  /export/gpx:
    get:
      tags: [Import Export]
      summary: Fahrten als GPX exportieren
      parameters:
        - { name: tripId, in: query, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: GPX-Datei
          content:
            application/gpx+xml:
              schema: { type: string, format: binary }
...
  /import/csv:
    post:
      tags: [Import Export]
      summary: Fahrten aus CSV importieren
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                dryRun: { type: boolean, default: false }
      responses:
        '200': { description: Importergebnis }
  /import/gpx:
    post:
      tags: [Import Export]
      summary: GPX importieren
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                vehicleId: { type: string, format: uuid }
      responses:
        '200': { description: Importergebnis }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...